refactor(cms): migrate ContentOverview onto the shared FilterOptions shell
Removes FilterOptionsDesktop.vue and FilterOptionsMobile.vue (161
lines); their content moves into the shared component's search /
extra-filters / extra-filters-mobile / selected-filters slots.
ContentOverview.vue's own props/model contract is unchanged. All
filter logic (status, translation, tags, sort, the min-length
Enter/Esc search) is untouched — only which component renders the
responsive shell changes.
